Python-API/requirements.txt
Flask==2.2.2
pymongo==4.2.0
python-dotenv==0.21.0
requests==2.28.2
pandas==1.5.3       
numpy==1.23.4       
scikit-learn==1.2.0 
flask-cors==3.0.10
